@@ -1,30 +1,38 @@
 package ngo.teog.swift.gui;
 
 import android.content.Context;
-import android.content.DialogInterface;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
 import android.text.Html;
+import android.view.MenuItem;
+
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
 
 import ngo.teog.swift.R;
 
 public abstract class BaseActivity extends AppCompatActivity {
-    public void showInfo(int stringID) {
+    public void showInfo(int stringId) {
         AlertDialog.Builder builder = new AlertDialog.Builder(this);
 
-        builder.setMessage(Html.fromHtml(getString(stringID)))
-                .setTitle("Information")
+        builder.setMessage(Html.fromHtml(getString(stringId)))
+                .setTitle(getString(R.string.info_dialog_heading))
                 .setIcon(R.drawable.ic_info_outline_black_24dp)
-                .setPositiveButton("Got it!", new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog, int id) {
-                //ignore
-            }
-        });
+                .setPositiveButton(getString(R.string.info_dialog_confirmation), (dialog, id) -> {
+                    //ignore
+                });
 
         AlertDialog dialog = builder.create();
         dialog.show();
+    }
+
+    public boolean onOptionsItemSelected(MenuItem item, int stringId) {
+        if(item.getItemId() == R.id.info) {
+            showInfo(stringId);
+            return true;
+        } else {
+            return super.onOptionsItemSelected(item);
+        }
     }
 
     public boolean checkForInternetConnection() {
@@ -33,11 +41,7 @@ public abstract class BaseActivity extends AppCompatActivity {
         if(cm != null) {
             NetworkInfo activeNetwork = cm.getActiveNetworkInfo();
 
-            if (activeNetwork != null && activeNetwork.isConnectedOrConnecting()) {
-                return true;
-            } else {
-                return false;
-            }
+            return activeNetwork != null && activeNetwork.isConnectedOrConnecting();
         } else {
             return false;
         }
