@@ -1,15 +1,11 @@
 package ngo.teog.swift.gui.main;
 
 import android.Manifest;
-import android.app.Dialog;
-import android.app.NotificationManager;
-import android.content.DialogInterface;
+import android.content.Context;
+import android.content.Intent;
+import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
-import android.support.v4.app.ActivityCompat;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AlertDialog;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -17,10 +13,13 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ProgressBar;
-import android.widget.TextView;
 import android.widget.Toast;
 
-import com.android.volley.RequestQueue;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProviders;
+
 import com.google.zxing.ResultPoint;
 import com.journeyapps.barcodescanner.BarcodeCallback;
 import com.journeyapps.barcodescanner.BarcodeResult;
@@ -28,12 +27,23 @@ import com.journeyapps.barcodescanner.DecoratedBarcodeView;
 
 import java.util.List;
 
-import ngo.teog.swift.R;
-import ngo.teog.swift.communication.RequestFactory;
-import ngo.teog.swift.communication.VolleyManager;
-import ngo.teog.swift.gui.BaseFragment;
+import javax.inject.Inject;
 
-public class BarcodeFragment extends BaseFragment {
+import ngo.teog.swift.R;
+import ngo.teog.swift.gui.deviceInfo.DeviceInfoActivity;
+import ngo.teog.swift.helpers.Defaults;
+import ngo.teog.swift.helpers.ResourceKeys;
+import ngo.teog.swift.helpers.data.AppModule;
+import ngo.teog.swift.helpers.data.DaggerAppComponent;
+import ngo.teog.swift.helpers.data.RoomModule;
+import ngo.teog.swift.helpers.data.ViewModelFactory;
+
+public class BarcodeFragment extends Fragment {
+
+    @Inject
+    ViewModelFactory viewModelFactory;
+
+    private BarcodeViewModel viewModel;
 
     private DecoratedBarcodeView barcodeScannerView;
     private String lastText;
@@ -55,7 +65,7 @@ public class BarcodeFragment extends BaseFragment {
             try {
                 int deviceNumber = Integer.parseInt(result.getText());
 
-                BarcodeFragment.this.invokeFetchRequest(deviceNumber);
+                invokeFetch(deviceNumber);
             } catch(NumberFormatException e) {
                 //ignore
             }
@@ -91,6 +101,14 @@ public class BarcodeFragment extends BaseFragment {
         if(ContextCompat.checkSelfPermission(this.getContext(), Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED) {
             ActivityCompat.requestPermissions(this.getActivity(), new String[]{Manifest.permission.CAMERA}, 0);
         }
+
+        DaggerAppComponent.builder()
+                .appModule(new AppModule(getActivity().getApplication()))
+                .roomModule(new RoomModule(getActivity().getApplication()))
+                .build()
+                .inject(this);
+
+        viewModel = ViewModelProviders.of(this, viewModelFactory).get(BarcodeViewModel.class);
     }
 
     @Override
@@ -106,29 +124,41 @@ public class BarcodeFragment extends BaseFragment {
         barcodeScannerView.pause();
     }
 
-    public void invokeFetchRequest(int id) {
-        RequestQueue queue = VolleyManager.getInstance(getActivity()).getRequestQueue();
-
-        RequestFactory.DefaultRequest request = new RequestFactory().createDeviceOpenRequest(getContext(), progressBar, searchButton, id);
-
-        searchField.setText(Integer.toString(id));
+    private void invokeFetch(int deviceId) {
+        searchField.setText(Integer.toString(deviceId));
         searchButton.setVisibility(View.INVISIBLE);
         progressBar.setVisibility(View.VISIBLE);
 
-        queue.add(request);
+        SharedPreferences preferences = this.getContext().getSharedPreferences(Defaults.PREF_FILE_KEY, Context.MODE_PRIVATE);
+        int userId = preferences.getInt(Defaults.ID_PREFERENCE, -1);
+
+        viewModel.init(userId, deviceId);
+        viewModel.getDeviceInfo().observe(BarcodeFragment.this, deviceInfo -> {
+            if(deviceInfo != null) {
+                Intent intent = new Intent(BarcodeFragment.this.getContext(), DeviceInfoActivity.class);
+                intent.putExtra(ResourceKeys.DEVICE_ID, deviceInfo.getDevice().getId());
+                BarcodeFragment.this.startActivity(intent);
+            } else {
+                Toast.makeText(this.getContext().getApplicationContext(), "device not found", Toast.LENGTH_SHORT).show();
+            }
+
+            searchField.setText(null);
+            searchButton.setVisibility(View.VISIBLE);
+            progressBar.setVisibility(View.INVISIBLE);
+        });
     }
 
     public boolean onKeyDown(int keyCode, KeyEvent event) {
         return barcodeScannerView.onKeyDown(keyCode, event);
     }
 
-    public void search() {
+    private void search() {
         String searchString = searchField.getText().toString();
 
         try {
             int deviceNumber = Integer.parseInt(searchString);
 
-            this.invokeFetchRequest(deviceNumber);
+            this.invokeFetch(deviceNumber);
         } catch(NumberFormatException e) {
             Toast.makeText(this.getContext().getApplicationContext(), "invalid device number", Toast.LENGTH_SHORT).show();
         }
