@@ -2,46 +2,52 @@ package ngo.teog.swift.gui.main;
 
 import android.content.Context;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.ListView;
-import android.widget.ProgressBar;
 import android.widget.TextView;
 
-import com.android.volley.RequestQueue;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProviders;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.concurrent.TimeUnit;
+import java.util.Collections;
+import java.util.Date;
+import java.util.List;
 
-import androidx.work.Constraints;
-import androidx.work.NetworkType;
-import androidx.work.PeriodicWorkRequest;
-import androidx.work.WorkManager;
-import ngo.teog.swift.gui.BaseFragment;
-import ngo.teog.swift.gui.DeviceInfoActivity;
+import javax.inject.Inject;
+
 import ngo.teog.swift.R;
-import ngo.teog.swift.communication.RequestFactory;
-import ngo.teog.swift.communication.VolleyManager;
-import ngo.teog.swift.helpers.HospitalDevice;
-import ngo.teog.swift.helpers.SearchObject;
+import ngo.teog.swift.gui.deviceInfo.DeviceInfoActivity;
+import ngo.teog.swift.helpers.Defaults;
 import ngo.teog.swift.helpers.DeviceState;
-import ngo.teog.swift.helpers.UpdateWorker;
+import ngo.teog.swift.helpers.ResourceKeys;
+import ngo.teog.swift.helpers.data.AppModule;
+import ngo.teog.swift.helpers.data.DaggerAppComponent;
+import ngo.teog.swift.helpers.data.DeviceInfo;
+import ngo.teog.swift.helpers.data.HospitalDevice;
+import ngo.teog.swift.helpers.data.Report;
+import ngo.teog.swift.helpers.data.ReportInfo;
+import ngo.teog.swift.helpers.data.RoomModule;
+import ngo.teog.swift.helpers.data.ViewModelFactory;
 
-public class TodoFragment extends BaseFragment {
+public class TodoFragment extends Fragment {
+
+    @Inject
+    ViewModelFactory viewModelFactory;
+
+    private TodoViewModel viewModel;
 
     private CustomSimpleArrayAdapter adapter;
-    private ListView listView;
-    private ProgressBar progressBar;
 
-    //TODO Die Konstante muss natürlich hier weg
-    public static final SimpleDateFormat DATE_FORMAT = new SimpleDateFormat("yyyy-MM-dd");
+    private List<DeviceInfo> values = new ArrayList<>();
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
@@ -50,36 +56,74 @@ public class TodoFragment extends BaseFragment {
 
     @Override
     public void onViewCreated(View view, Bundle savedInstanceState) {
-        listView = view.findViewById(R.id.maintenanceList);
-        ArrayList<SearchObject> values = new ArrayList<>();
-
-        progressBar = view.findViewById(R.id.progressBar);
+        ListView listView = view.findViewById(R.id.maintenanceList);
 
         adapter = new CustomSimpleArrayAdapter(getContext(), values);
         listView.setAdapter(adapter);
 
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-                Intent intent = new Intent(getContext(), DeviceInfoActivity.class);
-                intent.putExtra("device", (HospitalDevice)adapterView.getItemAtPosition(i));
-                startActivity(intent);
-            }
+        listView.setOnItemClickListener((adapterView, view1, i, l) -> {
+            Intent intent = new Intent(getContext(), DeviceInfoActivity.class);
+            intent.putExtra(ResourceKeys.DEVICE_ID, ((DeviceInfo)adapterView.getItemAtPosition(i)).getDevice().getId());
+            startActivity(intent);
         });
 
         final SwipeRefreshLayout swipeRefreshLayout = view.findViewById(R.id.swiperefresh);
 
         swipeRefreshLayout.setOnRefreshListener(
-            new SwipeRefreshLayout.OnRefreshListener() {
-                @Override
-                public void onRefresh() {
+                () -> {
                     swipeRefreshLayout.setRefreshing(false);
                     refresh();
                 }
-            }
         );
 
-        refresh();
+        DaggerAppComponent.builder()
+                .appModule(new AppModule(getActivity().getApplication()))
+                .roomModule(new RoomModule(getActivity().getApplication()))
+                .build()
+                .inject(this);
+
+        SharedPreferences preferences = this.getContext().getSharedPreferences(Defaults.PREF_FILE_KEY, Context.MODE_PRIVATE);
+        int id = preferences.getInt(Defaults.ID_PREFERENCE, -1);
+
+        viewModel = ViewModelProviders.of(this, viewModelFactory).get(TodoViewModel.class);
+        viewModel.init(id);
+        viewModel.getDeviceInfos().observe(this, deviceInfos -> {
+            if(deviceInfos != null && deviceInfos.size() > 0) {
+                this.values = deviceInfos;
+                adapter.clear();
+
+                for(DeviceInfo deviceInfo : deviceInfos) {
+                    List<ReportInfo> reports = deviceInfo.getReports();
+                    Collections.reverse(reports);
+                }
+
+                Collections.sort(deviceInfos, (first, second) -> {
+                    List<ReportInfo> firstReports = first.getReports();
+                    List<ReportInfo> secondReports = second.getReports();
+
+                    if(firstReports.size() > 0 && secondReports.size() > 0) {
+                        int firstState = firstReports.get(0).getReport().getCurrentState();
+                        int secondState = secondReports.get(0).getReport().getCurrentState();
+
+                        return (firstState-secondState)*-1;
+                    } else {
+                        return 0;
+                    }
+                });
+
+                for(DeviceInfo deviceInfo : deviceInfos) {
+                    List<ReportInfo> reportInfos = deviceInfo.getReports();
+
+                    if(reportInfos.size() > 0) {
+                        int currentState = reportInfos.get(0).getReport().getCurrentState();
+
+                        if(currentState == 1 || currentState == 2 || currentState == 3) {
+                            adapter.add(deviceInfo);
+                        }
+                    }
+                }
+            }
+        });
     }
 
     @Override
@@ -90,45 +134,22 @@ public class TodoFragment extends BaseFragment {
     }
 
     private void refresh() {
-        Constraints updateConstraints = new Constraints.Builder()
-                .setRequiredNetworkType(NetworkType.CONNECTED)
-                .build();
+        SharedPreferences preferences = this.getContext().getSharedPreferences(Defaults.PREF_FILE_KEY, Context.MODE_PRIVATE);
+        int userId = preferences.getInt(Defaults.ID_PREFERENCE, -1);
 
-        WorkManager.getInstance().cancelAllWorkByTag("update_todo");
-
-        PeriodicWorkRequest updateWork = new PeriodicWorkRequest.Builder(UpdateWorker.class, 6, TimeUnit.HOURS)
-                .addTag("update_todo")
-                .setConstraints(updateConstraints)
-                .build();
-
-        WorkManager.getInstance().enqueue(updateWork);
-
-        progressBar.setVisibility(View.VISIBLE);
-        listView.setVisibility(View.INVISIBLE);
-
-        if(this.checkForInternetConnection()) {
-            RequestQueue queue = VolleyManager.getInstance(getContext()).getRequestQueue();
-
-            RequestFactory.DeviceListRequest request = new RequestFactory().createTodoListRequest(getContext(), progressBar, listView, adapter);
-
-            queue.add(request);
-        }
+        viewModel.refreshHospital(userId);
     }
 
-    private class CustomSimpleArrayAdapter extends ArrayAdapter<SearchObject> {
-        private final Context context;
-
-        private CustomSimpleArrayAdapter(Context context, ArrayList<SearchObject> values) {
+    private class CustomSimpleArrayAdapter extends ArrayAdapter<DeviceInfo> {
+        private CustomSimpleArrayAdapter(Context context, List<DeviceInfo> values) {
             super(context, -1, values);
-            this.context = context;
         }
 
-
-
         @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
+        @NonNull
+        public View getView(int position, View convertView, @NonNull ViewGroup parent) {
             if(convertView == null) {
-                LayoutInflater inflater = (LayoutInflater) context
+                LayoutInflater inflater = (LayoutInflater) this.getContext()
                         .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
                 convertView = inflater.inflate(R.layout.row_todo, parent, false);
             }
@@ -137,16 +158,24 @@ public class TodoFragment extends BaseFragment {
             TextView dateView = convertView.findViewById(R.id.dateView);
             TextView statusView = convertView.findViewById(R.id.statusView);
             ImageView imageView = convertView.findViewById(R.id.imageView);
+            TextView detailView = convertView.findViewById(R.id.detailView);
 
-            HospitalDevice device = (HospitalDevice)this.getItem(position);
+            DeviceInfo deviceInfo = this.getItem(position);
 
-            if(device != null) {
+            if(deviceInfo != null) {
+                HospitalDevice device = deviceInfo.getDevice();
+                Report lastReport = deviceInfo.getReports().get(0).getReport();
+
                 nameView.setText(device.getType());
 
-                String dateString = DATE_FORMAT.format(device.getLastReportDate());
+                long now = new Date().getTime();
+                Date reportDate = lastReport.getCreated();
+                String dateString = (now-reportDate.getTime())/1000/60/60/24 + " d";
                 dateView.setText(dateString);
 
-                DeviceState triple = DeviceState.buildState(device.getState(),this.getContext());
+                detailView.setText(device.getManufacturer() + "\n" + device.getModel());
+
+                DeviceState triple = DeviceState.buildState(lastReport.getCurrentState(), this.getContext());
 
                 statusView.setText(device.getWard());
                 imageView.setImageDrawable(triple.getStateicon());
