@@ -7,14 +7,14 @@ import ngo.teog.swift.R;
 
 public class DeviceState {
 
-    private static final int ICONS[] = {R.drawable.ic_check,
+    private static final int[] ICONS= {R.drawable.ic_check,
                                         R.drawable.ic_maintenance,
                                         R.drawable.ic_repair,
                                         R.drawable.ic_in_progress,
                                         R.drawable.ic_broken_salvage,
                                         R.drawable.ic_working_with_limitations};
 
-    private static final int COLORS[] = {android.R.color.holo_green_dark,
+    private static final int[] COLORS = {android.R.color.holo_green_dark,
                                         android.R.color.holo_blue_light,
                                         android.R.color.holo_orange_dark,
                                         android.R.color.holo_green_light,
